Move new Asset/ComponentDataImpl classes
Moves the new builder-backed `AssetDataImpl`
and `ComponentDataImpl`classes from `datalayer.builder`
package to `datalayer` package.
